refactor: rewrite cjsToEsm module to ESM syntax
